c_ui: use linux power_supply subsystem for battery
Use the mainline Linux power_supply subsystem for getting information
about battery voltage and current. There is no known hardware that can
measure motor current, so that was removed (it was only a feature in
pre-release EV3 hardware).

The first "System" power supply is used. If not present, the battery
voltage is faked at 9V. This allows us to re-enable the low voltage
shutdown option.

Closes #11