Remove preempted vm option
This option is not used in the official firmware. It relies on a timer
in the d_analog driver that we don't want to depend on either. Since
it is unlikely that we will ever use this, let's just get it out of the
way to make the code cleaner.