c_com: Replace wpa_supplicant with ConnMan
c_wifi was using private wpa_supplicant headers to make calls into that
library, so the code here was not very workable anyway.

ev3dev uses ConnMan for network management, so that is what we are going
to use here. We are also going to included wired etherent in the connections
since BeagleBone and Raspberry Pi have built-in ethernet.

This is not complete. So far, getting network connection info (IP/MAC) is
working. Scanning and listing APs is working.

Also cleaned up c_wifi debug prints and added a compiler option since there
was not one already.