c_com: get WiFi communications working again.
Re-implemented the TCP and UDP socket code to use GIO. Modified the read
and write functions to use GIO streams. There is no longer a need for the
wifi state machine as everything is handled via callbacks in the glib main
loop now.

The re-implementation of the socket code includes a fix to propertly take
into account the netmask for the UDP broadcast socket.

Closes #4