c_com: Fix blocking reads in c_wifi
Apparently g_output_stream_write/g_input_stream_read block even when the
underlying socket is set to non-blocking. Because of the way the exiting
code is structured, we can't use the aysnc versions of these funtions. So,
we have to use the lower level g_socket_send/g_socket receive instead.