c_com: Manage usb hid systemd service
Originally, the idea was to bring up the usb hid gadget using a udev
rule. However, it turns out that the EV3-G software still detects this
even if lms2012 is not running, which causes communication errors. So,
we are now starting the service using systemd dbus within lms2012 so that
the usb is only configured while lms2012 is running.

This also helps if/when the usb hid gadget driver gets into a bad state.
Restarting lms2012 will tear down the driver and bring it up again.

Fixes #27